requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
matplotlib
jupyter
